moved protocol beans from MifosChatbotApplication to ProtocolConfig
