+++ server/src/main/java/org/mifos/chatbot/server/MifosChatbotApplication.java
@@ -15,13 +15,9 @@
  */
 package org.mifos.chatbot.server;
 
-import com.github.messenger4j.Messenger;
-import org.mifos.chatbot.protocol.telegram.TelegramChatService;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.cache.annotation.EnableCaching;
-import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.EnableAspectJAutoProxy;
 import org.telegram.telegrambots.ApiContextInitializer;
@@ -31,18 +27,6 @@ import org.telegram.telegrambots.ApiContextInitializer;
 @EnableCaching(proxyTargetClass = true)
 @EnableAspectJAutoProxy(proxyTargetClass = true)
 public class MifosChatbotApplication {
-    @Bean
-    public Messenger messenger(@Value("${messenger4j.pageAccessToken}") String pageAccessToken,
-                               @Value("${messenger4j.appSecret}") final String appSecret,
-                               @Value("${messenger4j.verifyToken}") final String verifyToken) {
-        return Messenger.create(pageAccessToken, appSecret, verifyToken);
-    }
-
-    @Bean
-    public TelegramChatService telegramChatService() {
-        return new TelegramChatService();
-    }
-
     public static void main(String[] args) {
         ApiContextInitializer.init();
         SpringApplication.run(MifosChatbotApplication.class, args);
