+++ server/src/main/java/org/mifos/chatbot/server/config/ProtocolConfig.java
@@ -15,6 +15,7 @@
  */
 package org.mifos.chatbot.server.config;
 
+import com.github.messenger4j.Messenger;
 import com.ullink.slack.simpleslackapi.SlackSession;
 import com.ullink.slack.simpleslackapi.impl.SlackSessionFactory;
 import lombok.extern.slf4j.Slf4j;
@@ -23,7 +24,9 @@ import org.mifos.chatbot.core.model.Message;
 import org.mifos.chatbot.core.model.MifosResponse;
 import org.mifos.chatbot.core.model.MifosSettings;
 import org.mifos.chatbot.protocol.slack.SlackChatService;
+import org.mifos.chatbot.protocol.telegram.TelegramChatService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
@@ -61,4 +64,17 @@ public class ProtocolConfig {
 
         return slackChatService;
     }
+
+    @Bean
+    public Messenger messenger(@Value("${messenger4j.pageAccessToken}") String pageAccessToken,
+                               @Value("${messenger4j.appSecret}") final String appSecret,
+                               @Value("${messenger4j.verifyToken}") final String verifyToken) {
+        return Messenger.create(pageAccessToken, appSecret, verifyToken);
+    }
+
+    @Bean
+    public TelegramChatService telegramChatService() {
+        return new TelegramChatService();
+    }
+
 }
